Move backend, add github actions, kill circleci
