+++ .github/workflows/main.yml
@@ -1,0 +1,26 @@
+name: CI
+
+# Events that trigger the workflow
+on: push
+
+jobs:
+  build-and-test:
+    runs-on: ubuntu-latest
+    steps:
+      # https://github.com/actions/checkout
+      - uses: actions/checkout@v2
+      - name: Run build
+        run: |
+          docker-compose build
+      - name: Test and generate coverage report
+        run: |
+          pip install fastapi
+          pip install pytest-cov
+          py.test --cov-report term --cov=./ --cov-report=xml
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v2
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          files: coverage.xml
+          env_vars: OS,PYTHON
+          fail_ci_if_error: true
