--- app/main.py
@@ -1,7 +1,0 @@
-def foo(bar: str) -> str:
-    string = f"foo{bar}"
-    print(string)
-    return string
-
-
-foo("bar")
