--- tests/test_main.py
@@ -1,9 +1,0 @@
-from app.main import foo
-
-
-def test_foo():
-    assert foo("bar") == "foobar"
-
-
-def test_false_foo():
-    assert foo("foo") != "foobar"
